requirements.txt
Django==2.0.5
django-ckeditor==5.4.0
django-js-asset==1.1.0
django-sitetree==1.10.0
django-solo==1.1.3
django-webpack-loader==0.6.0
Pillow==5.1.0
pytz==2018.4
Unidecode==1.0.22
